day4/requirements.txt
openai>=1.30.0
anthropic>=0.34.0
python-dotenv>=1.0.1
